+++ src/at/terranova/generation/biomes/SavannaBiome.java
@@ -16,16 +16,16 @@ import java.util.Random;
 
 public class SavannaBiome implements CustomBiome {
     @Override
-    public void generate(NoiseProvider provider, WorldInfo worldInfo, Random random, int x, double height, int z, ChunkGenerator.ChunkData chunkData) {
+    public void generate(NoiseProvider provider, WorldInfo worldInfo, Random random, int x, int height, int z, ChunkGenerator.ChunkData chunkData) {
         int dirtDepth = 8 + random.nextInt(3);
-        for(int i = (int) height - 1; i > height - dirtDepth; i--) {
+        for(int i = height - 1; i > height - dirtDepth; i--) {
             chunkData.setBlock(x, i, z, Material.DIRT);
         }
-        for (int i = (int) height - dirtDepth; i > 0; i--) {
+        for (int i =height - dirtDepth; i > 0; i--) {
             chunkData.setBlock(x, i, z, Material.STONE);
         }
 
-        chunkData.setBlock(x, (int) height, z, Material.GRASS_BLOCK);
+        chunkData.setBlock(x, height, z, Material.GRASS_BLOCK);
     }
 
     @Override
